Создал сервисную функцию, которая будет отвечать за выборку категорий и которую можно переиспользовать в любом месте системы.
